# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectData = { [key: string]: unknown };

export type ProjectRole = typeof ProjectRole[keyof typeof ProjectRole];


export const ProjectRole = {
  owner: 'owner',
  planner: 'planner',
  family: 'family',
  viewer: 'viewer',
} as const;

export interface Project {
  id: string;
  title: string;
  data: ProjectData;
  role: ProjectRole;
  updatedAt: string;
}

export type ProjectInputData = { [key: string]: unknown };

export interface ProjectInput {
  /**
     * @minLength 1
     * @maxLength 160
     */
  title: string;
  data: ProjectInputData;
}

export type ProjectUpdate = ProjectInput & {
  updatedAt: string;
};

export const DeleteConfirmationValue = {
  confirmation: 'SUPPRIMER',
} as const;
export type DeleteConfirmation = typeof DeleteConfirmationValue;

export type InvitationInputRole = typeof InvitationInputRole[keyof typeof InvitationInputRole];


export const InvitationInputRole = {
  planner: 'planner',
  family: 'family',
  viewer: 'viewer',
} as const;

export interface InvitationInput {
  email: string;
  role: InvitationInputRole;
}

export interface UploadInput {
  projectId: string;
  /**
     * @minLength 1
     * @maxLength 240
     */
  name: string;
  contentType: string;
  /**
     * @minimum 1
     * @maximum 26214400
     */
  size: number;
}

export interface UploadTicket {
  uploadURL: string;
  objectPath: string;
}

export type RsvpInputStatus = typeof RsvpInputStatus[keyof typeof RsvpInputStatus];


export const RsvpInputStatus = {
  confirmed: 'confirmed',
  declined: 'declined',
} as const;

export type RsvpInputAttendance = {
  ceremony: boolean;
  cocktail: boolean;
  dinner: boolean;
  brunch: boolean;
};

export interface RsvpInput {
  status: RsvpInputStatus;
  attendance: RsvpInputAttendance;
  /** @maxLength 1000 */
  dietary?: string;
  plusOne: boolean;
  /** @maxLength 2000 */
  notes?: string;
}