# lib/api-zod/src/generated/types/deleteConfirmation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export const DeleteConfirmationValue = {
  confirmation: 'SUPPRIMER',
} as const;
export type DeleteConfirmation = typeof DeleteConfirmationValue;