# lib/api-zod/src/generated/types/invitationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InvitationInputRole } from './invitationInputRole';

export interface InvitationInput {
  email: string;
  role: InvitationInputRole;
}