# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectData } from './projectData';
import type { ProjectRole } from './projectRole';

export interface Project {
  id: string;
  title: string;
  data: ProjectData;
  role: ProjectRole;
  updatedAt: Date;
}