# lib/api-zod/src/generated/types/projectInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectInputData } from './projectInputData';

export interface ProjectInput {
  /**
     * @minLength 1
     * @maxLength 160
     */
  title: string;
  data: ProjectInputData;
}