# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectInput } from './projectInput';

export type ProjectUpdate = ProjectInput & {
  updatedAt: Date;
};