# lib/api-zod/src/generated/types/rsvpInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RsvpInputAttendance } from './rsvpInputAttendance';
import type { RsvpInputStatus } from './rsvpInputStatus';

export interface RsvpInput {
  status: RsvpInputStatus;
  attendance: RsvpInputAttendance;
  /** @maxLength 1000 */
  dietary?: string;
  plusOne: boolean;
  /** @maxLength 2000 */
  notes?: string;
}