# lib/api-zod/src/generated/types/uploadInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UploadInput {
  projectId: string;
  /**
     * @minLength 1
     * @maxLength 240
     */
  name: string;
  contentType: string;
  /**
     * @minimum 1
     * @maximum 26214400
     */
  size: number;
}